Improvements for TIFF -> JPEG2000 conversion
Fixes an issue where if the TIFF had XMP values that conflicted with the
IFD0 values, the XMP and IFD0 could become conflated in the final image.
The new code saves the descriptive metadata (Artist, Make and Model)
from the XMP if needed, strips the XMP, and then reinserts the
descriptive metadata in the final image.