Remove explicit locking for feed queue
The whole table lock caused poor performance.

UPDATES to innodb tables should lock the rows that are in the process of
being updated. Also removes transactions from the SELECT-only queries.

I also added a index to (node,status) to minimize the number of rows
that get scanned in normal process, which should help as well.