HT-1746: preserve metadata from lossless JPEG2000 images
This was previously relying on the pathway for compressing incoming
TIFFs to JPEG2000 images, but that method had a lot of extra stuff to
deal with potentially bad metadata in the original TIFF. This now just
immediately recompresses the incoming JPEG2000 back to JPEG2000 and
copies all the metadata (via exiftool) from the original JPEG2000. It
then falls through to the existing jpeg2000 remediation to fix any
remaining metadata issues with the JPEG2000 image.