Add mariadb service for docker-compose
collate tests were running into issues with SQL statements that were not
supported with sqlite. This adds a docker-compose service for mariadb,
seeds it with the rights codes, and adds a setup for loading some
fixture data required for the tests.

Further work could wrap each test in a transaction and load the fixtures
separately for each test (or at least delete the fixtures at the end of
the test suite)

Also adds "wait-for" script to wait for mariadb to be up before running
tests via travis, since the database must be up for tests to pass.