Refactor collate operations
This is a start to a potential pattern for factoring out operations in
other stages. There is probably more that could be extracted in terms of
logging operations.

Note that set_error raises an exception (unless stop_on_error = 0 is set
in configuration), so if a sub-task doesn't complete the entire stage is
aborted.