HT-2433: storage backend for versioned pairtree
This adds a new storage backend for a "versioned pairtree" (basically a
pairtree where each object has versions stored in the object area) where
the versions are identified by a timestamp.

Versions are recorded in the new "feed_backups" table.

This factors out much of the functionality in Collate.pm to a new
HTFeed::Storage object that represents a particular storage backend.