HT-2433: post-move validation; refactor storage / collate
- Add post-move validation using the same process as pre-move validation

- Still to do: move the existing version aside so that we can roll back
if needed following post-validation

- HTFeed::Stage::Collate is responsible for flow of control

- HTFeed::Storage logs errors and retain them for inspection by the
stage and further reporting. This removes the need for HTFeed::Storage
and its subclasses to have a reference to the running stage.

It seems like there could still be a separation between the storage
backend itself and the common operations in the superclass that might be
better expressed by composition than inheritance. Worth thinking about
what from HTFeed::Volume should move to HTFeed::Storage, and in that
case which of the move/copy/validate operations should move out of
HTFeed::Storage.