HT-2433: verify METS against original copy
This protects against a case where the METS is corrupted in some way
such that it is still valid and usable for verifying the zip, but
doesn't match the originally-written version.

It's hard to think of a plausible case where the METS would still be
valid and have the correct ZIP checksum, but if we're verifying the zip
it seems to make sense to verify the METS as well.