Add storage migration stage
Storage migration allows copying the item from the main repository to other
configured storages, i.e. different backup locations.

Storage migration is essentially the same as collate, but looks at a
different configuration key for the storage classes to use. This also
adds a stage map for the storage migration.