HT-2502: StorageMigrate wasn't copying from repository
The storage_migrate.t test was previously succeeding because it was
copying from a staging area, not from the repository. This commit
adjusts the test so that it cleans out the staging area, then adjusts
the HTFeed::PackageType::HathiTrust::Volume to provide the path to the
mets and zip in the repository. It also adjusts TempDirs to properly use
temporary directories for the repository locations.