HT-2502: Address issues around StorageMigrate and cleaning
Previously, the behavior of HTFeed::StorageMigrate#clean was untested;
it is called by the job runner but isn't called in the context of
running a particular stage. As things were, this lead to the METS being
removed from the repository after running StorageMigrate.

A second bug with cleaning caused this NOT to happen on success. The
base collate stage attempted to call HTFeed::Volume#clean_sip_success,
which failed for the 'ht' package type. This led cleaning to fail before
it had a chance to remove the METS file.

This addresses the cleaning problems in two ways - one, it overrides the
cleaning behavior in HTFeed::Stage::StorageMigrate to avoid any cleaning
behavior on success or failure, while also testing that the stage does
not leave temporary files behind in staging directories. It also
overrides the cleaning behavior in the
HTFeed::PackageType::HathiTrust::Volume. (This split in cleaning
responsibility doesn't make much sense, but such is the current state of
the current code.) This area is ripe for refactoring when we move to a
job runner such that the cleaning should happen as an integral part of
the stage. Either fix is sufficient to address the problem, but the
belt-and-suspenders approach to help ensure a similar problem doesn't
occur in the future seems worthwhile.

This also addresses the underlying cause of at least one failure, which is
that some old zips in the repository had files such as 'notes.txt' that
were not present in the METS files. Storage migration should not fail
because of this (although we may want to audit for volumes like this at
some point in the future)