Mock ClamAV for tests
clamav docker image was failing with issues downloading virus
definitions. No compelling reason for CI to be dependent on this -- it
might be nice to add an integration test for this, but if we do, we
should try to remove the external dependency -- or at least allow the
rest of the test suite to run.