HT-3375: run clamscan on demand
Given the low number of files we need to scan (only items coming from
EMMA), running clamav in a sidecar container for every ingest worker is
a waste of resources; especially, downloading virus definitions for
every separate pod. This assumes virus definitions will be present at
/var/lib/clamav and calls out to clamscan when needed. It preserves the
same interface as ClamAV::Client, so switching back should be easy if
needed in the future.