Decompress before stripping XMP
There were some images that imagemagick can remediate but that exiftool
doesn't want to handle. As far as I can tell, there's no reason to
strip the XMP before decompressing, and in fact in testing it seems like
ImageMagick strips out the XMP anyway. I can't rule out that there won't
be some other image that will fail with this order of operations, but
generally this should be harmless.