GS-5046: Layers? It should have them.
There isn't any particular reason that we know of to enforce a
particular number of layers. We have seen a variety over the years; in
this case there are 3. Rather than enumerating another possible option
for something that isn't terribly important, just allow any value in the
range we've seen.

This is more important for Google-scanned material where a change in the
number of layers could reflect some other change in their image
compression process, but I don't think it should matter here.