ETT-571: handle empty DateTime in RGB TIFFs
"copy_old_to_new" was checking that the field existed previously, but
not that it was non-empty. This doesn't cover every possible scenario,
and there are likely other cases of invalid metadata in RGB TIFFs that
we may need to handle in the future, but we can deal with it when it
comes up.